# src/main/resources/META-INF/resources/openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: Order-Service API
  description: An API providing basic ordering capabilities
servers:
  - description: Local Order-Service API
    url: https://localhost:8080
tags:
  - name: orders
    description: Available ordering operations
paths:
  /orders:
    get:
      tags:
        - orders
      summary: Get all available orders
      operationId: getAllOrders
      responses:
        '200':
          description: OK. Collection of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      tags:
        - orders
      summary: Create a new order
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewOrder'
      responses:
        '200':
          description: OK. Order was successfully created
        '400':
          description: Invalid Input. The request payload contains invalid data
        '406':
          description: Not acceptable. Payment was not authorized
  /orders/{orderId}:
    get:
      tags:
        - orders
      summary: Get an order by order id
      operationId: getOrder
      parameters:
        - name: orderId
          in: path
          description: Order ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK. Order Details were successfully returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not found. Order was not found
    put:
      tags:
        - orders
      summary: Update Card of an Order
      operationId: updateCardOfOrder
      parameters:
        - name: orderId
          in: path
          description: Order ID
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Card'
      responses:
        '200':
          description: OK. Card Information was successfully updated
        '404':
          description: Not found. Order was not found
    delete:
      tags:
        - orders
      summary: Delete an order
      operationId: deleteOrder
      parameters:
        - name: orderId
          in: path
          description: Order ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK. The Order was successfully deleted
        '404':
          description: Order was not found
  /orders/{orderId}/items:
    get:
      tags:
        - orders
      summary: Get all items of an order
      operationId: getAllItemsFromOrder
      parameters:
        - name: orderId
          in: path
          description: Order ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK. Collection of items from an order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '404':
          description: Not found. Order was not found
  /orders/{orderId}/items/{itemId}:
    post:
      tags:
        - orders
      summary: Add an item to an order
      operationId: addItemToOrder
      parameters:
        - name: orderId
          in: path
          description: Order ID
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          description: Item ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK. A new item was added to the order
        '204':
          description: No content. Item already is included in the order
        '404':
          description: Not found. Order was not found
    get:
      tags:
        - orders
      summary: Get the item from an order
      operationId: getItemFromOrder
      parameters:
        - name: orderId
          in: path
          description: Order ID
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          description: Item ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK. Retrieved item was returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Not found. Order or Item were not found
    delete:
      tags:
        - orders
      summary: Delete an item from an order
      operationId: deleteItemFromOrder
      parameters:
        - name: orderId
          in: path
          description: Order ID
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          description: Item ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK. Item was successfully deleted
        '404':
          description: Not found. Order or Item were not found
components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: integer
          example: 1
        customer:
          $ref: '#/components/schemas/Customer'
        address:
          $ref: '#/components/schemas/Address'
        card:
          $ref: '#/components/schemas/Card'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        date:
          type: string
          format: date
        total:
          type: number
          example: 125.33
    Customer:
      type: object
      properties:
        id:
          type: integer
          example: 4
        firstName:
          type: string
          example: Max
        lastName:
          type: string
          example: Mustermann
        userName:
          type: string
          example: Username
    Card:
      type: object
      properties:
        id:
          type: integer
          example: 3
        longNum:
          type: string
          example: 1234-5678-9012-3456
        expires:
          type: string
          example: 08/89
        ccv:
          type: string
          example: 123
    Item:
      type: object
      properties:
        id:
          type: integer
          example: 2
        name:
          type: string
          example: TV-Screen
        quantity:
          type: integer
          example: 20
        unitPrice:
          type: number
          example: 20.50
    Address:
      type: object
      properties:
        id:
          type: integer
          example: 1
        number:
          type: string
          example: 111
        street:
          type: string
          example: Example Street
        city:
          type: string
          example: Munich
        postcode:
          type: string
          example: 80689
        country:
          type: string
          example: Germany
    NewOrder:
      type: object
      properties:
        customer:
          type: string
          format: uri
        address:
          type: string
          format: uri
        card:
          type: string
          format: uri
        items:
          type: string
          format: uri